int -> short.
git-svn-id: svn+ssh://svn.java.net/grizzly~svn/branches/2dot0/code@6035 04fdb32c-e91d-0410-84f5-98e489d73e21

+++ modules/grizzly/src/main/java/org/glassfish/grizzly/nio/NIOConnection.java
@@ -648,7 +648,7 @@ public abstract class NIOConnection implements Connection<SocketAddress> {
     protected final void checkEmptyRead(final int size) {
         if (WIN32) {
             if (size == 0) {
-                final int count = ++zeroByteReadCount;
+                final short count = ++zeroByteReadCount;
                 if (count >= MAX_ZERO_READ_COUNT) {
                     try {
                         close();
